% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromRxode2.R
\name{.setRxode2}
\alias{.setRxode2}
\title{This tells 'rxode2et' that 'rxode2' is available}
\usage{
.setRxode2()
}
\value{
Nothing called for side effects
}
\description{
This really should not be called by a user.  This call allows
'rxode2' solves to be a bit more flexible (and actually work!)
}
\author{
Matthew L. Fidler
}
\keyword{internal}
