% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R, R/etTran.R
\name{is.rxEt}
\alias{is.rxEt}
\title{Check to see if this is an rxEt object.}
\usage{
is.rxEt(x)

is.rxEt(x)
}
\arguments{
\item{x}{object to check to see if it is rxEt

If this is an rxEt object that has expired strip all rxEt
information.}
}
\value{
Boolean indicating if this is a rxode2 event table

Boolean indicating if this is a rxode2 event table
}
\description{
Check to see if this is an rxEt object.

Check to see if this is an rxEt object.
}
\author{
Matthew L.Fidler
}
\keyword{internal}
