% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{lotri}
\alias{rxRmvn}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lotri}{\code{\link[lotri]{lotri}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{rxode2random}{\code{\link[rxode2random]{rxRmvn}}}
}}

\value{ Inherited from parent routine }
